--- src/main/java/br/com/cadastros/cadastrausuario/application/api/UsuarioResponse.java
@@ -1,11 +1,0 @@
-package br.com.cadastros.cadastrausuario.application.api;
-
-import lombok.Builder;
-import lombok.Value;
-
-import java.util.UUID;
-@Value
-@Builder
-public class UsuarioResponse {
-    private UUID idUsuario;
-}
