+++ src/main/java/br/com/cadastros/cadastrausuario/application/parentes/application/api/ParenteDetailResponse.java
@@ -1,0 +1,19 @@
+package br.com.cadastros.cadastrausuario.application.parentes.application.api;
+
+import lombok.Value;
+
+import java.util.UUID;
+
+@Value
+public class ParenteDetailResponse {
+
+    private UUID idParente;
+    private String nome;
+    private String sobrenome;
+
+    public ParenteDetailResponse(String nome, String cpf) {
+        this.nome = getNome();
+        this.idParente = getIdParente();
+        this.sobrenome = getSobrenome();
+    }
+}
