+++ src/main/java/br/com/cadastros/cadastrausuario/application/parentes/infra/ParenteSpringDataJPARepository.java
@@ -1,6 +1,6 @@
 package br.com.cadastros.cadastrausuario.application.parentes.infra;
 
-import br.com.cadastros.cadastrausuario.application.domain.Parente;
+import br.com.cadastros.cadastrausuario.application.parentes.domain.Parente;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
@@ -8,5 +8,5 @@ import java.util.UUID;
 
 public interface ParenteSpringDataJPARepository extends JpaRepository<Parente, UUID> {
     Parente save(Parente parente);
-    List<Parente> findAllByIdUsuario(UUID idUsuario);
+    List<Parente> findParenteByIdUsuario(UUID idUsuario);
 }
