--- src/main/java/br/com/cadastros/cadastrausuario/application/repository/UsuarioRepository.java
@@ -1,14 +1,0 @@
-package br.com.cadastros.cadastrausuario.application.repository;
-
-import br.com.cadastros.cadastrausuario.application.domain.Parente;
-import br.com.cadastros.cadastrausuario.application.domain.Usuario;
-
-import java.util.List;
-
-public interface UsuarioRepository {
-
-    Usuario save(Usuario usuario);
-    List<Usuario> buscaUsuarios();
-    Usuario buscarUsuarioPorCpf(String cpf);
-    void deleteUsuario(Usuario usuario);
-}
